修复了内存泄漏bug，修改了Info类，进行了xmlparse测试，修改了xmlParseType, 修复了B+树数据块bug，添加了Dblp_bpt_ms

+++ DS_2020/DS_2020_Shared/BPlusTreePool.h
@@ -17,7 +17,7 @@ namespace DS_BPlusTree {
             , _block_size(_block_size)
             , _page_num(_page_num)
         {
-            _page_size = ((size_t)_block_size - PAGE_T::getSize() + sizeof(PAGE_T));
+            _page_size = (unsigned int) ((size_t)_block_size - PAGE_T::getSize() + sizeof(PAGE_T));
             size_t sz = _page_size * _page_num;
             caches = malloc(sz);
             memset(caches, 0, sz);
@@ -182,7 +182,7 @@ namespace DS_BPlusTree {
             , _block_size(_block_size)
             , _page_num(_page_num)
         {
-            _page_size = ((size_t)_block_size - PAGE_T::getSize() + sizeof(PAGE_T));
+            _page_size = (unsigned int)((size_t)_block_size - PAGE_T::getSize() + sizeof(PAGE_T));
             size_t sz = _page_size * _page_num;
             caches = malloc(sz);
             memset(caches, 0, sz);
